requirements.txt
# Core MCP dependencies
mcp>=1.0.0
fastmcp>=0.8.0

# LangChain dependencies
langchain-openai>=0.1.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0

# Additional dependencies that may be required
asyncio-contextmanager>=1.0.0
pydantic>=2.0.0
starlette>=0.27.0
fastapi[standard]>=0.104.0

# Optional: For better logging and development
python-dotenv>=1.0.0